feat: sidebar for the explore page
