+++ ui/src/components/index.ts
@@ -3,6 +3,8 @@ export { default as HorizontalNavbar } from './HorizontalNavbar.svelte';
 export { default as InlineItem } from './InlineItem.svelte';
 export { default as InlineShip } from './InlineShip.svelte';
 export { default as MobileHeader } from './MobileHeader.svelte';
+export { default as RadioStations } from './RadioStations.svelte';
+export { default as TopCreators } from './TopCreators.svelte';
 export { default as VerticalNavbar } from './VerticalNavbar.svelte';
 
 export { default as AppCard } from './AppCard.svelte';
